+++ app/router.js
@@ -4,15 +4,14 @@
  * Router
  * Jesús Juan Aguilar 2017
  * */
-//const datapool = require('./datapool');
+const datapool = require('./datapool');
 
 function defaultRedirect(req, res) {
   res.redirect('/');
 }
 
 module.exports = function router(app) {
-  //datapool.loadRepositories();
-  //datapool.loadBussinesOperations(app);
+  datapool.loadEnviroment(app);
   app.get('*', defaultRedirect);
 };