+++ src/routes.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import { Route, IndexRoute } from 'react-router'
 import Layout from './components/Layout';
-import IndexPage from './components/IndexPage';
-import AthletePage from './components/AthletePage';
 import NotFoundPage from './components/NotFoundPage';
 import GridComponent from './components/GridComponent';