+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## Ng2 login customized
 
 ### X.X.X | New Version
+~ Application Notifications [https://www.npmjs.com/package/kynplex-react-notifications]
+
+### 0.0.1 | 14/04/2017
 * Default React project structure
 * Webpack configuration development
 * Login view
-~ Login crud
-~ Ajax library
-~ Adding redux
+* Adding redux
+* Ajax library
+* User, Token BLL
+* Login Backend
+* Login Signup